+++ tests/testthat/test-gs4_auth.R
@@ -57,3 +57,37 @@ test_that("gs4_auth_configure(app =) is deprecated in favor of client", {
   expect_equal(gs4_oauth_client()$id, "abc.apps.googleusercontent.com")
 })
 
+# gs4_scopes() ----
+test_that("gs4_scopes() reveals Sheets scopes", {
+  expect_snapshot(gs4_scopes())
+})
+
+test_that("gs4_scopes() substitutes actual scope for short form", {
+  expect_equal(
+    gs4_scopes(c(
+      "spreadsheets",
+      "drive",
+      "drive.readonly"
+    )),
+    c(
+      "https://www.googleapis.com/auth/spreadsheets",
+      "https://www.googleapis.com/auth/drive",
+      "https://www.googleapis.com/auth/drive.readonly"
+    )
+  )
+})
+
+test_that("gs4_scopes() passes unrecognized scopes through", {
+  expect_equal(
+    gs4_scopes(c(
+      "email",
+      "spreadsheets.readonly",
+      "https://www.googleapis.com/auth/cloud-platform"
+    )),
+    c(
+      "email",
+      "https://www.googleapis.com/auth/spreadsheets.readonly",
+      "https://www.googleapis.com/auth/cloud-platform"
+    )
+  )
+})